app/requirements.txt
pyyaml
hvac
requests
pyhcl